Stop context menu protruding above viewport, allow nested context menus
When the menu is too big it is set to display above the x,y where the mouse click was which can put the top of the menu above the viewport (unreachable).  This disables offsetting the menu if it's height is greater than the window.

If the innermost menu is disabled when declaring context menus on nested elements the event doesn't bubble up to the outer menus.  This contains a first attempt at fixing that issue, however proper test cases need to be constructed to solve this issue.